Add a NexusClient for getting legacy tokens
The only method supported is getting a legacy token. To make this work cleanly,
also added a passthrough to allow more kwargs to come through BaseClient
methods into requests.
Removed the relevant functionality from the TransferClient class.
