+++ globus_sdk/__init__.py
@@ -1,3 +1,4 @@
 
 from globus_sdk.base import GlobusError, GlobusResponse
 from globus_sdk.transfer import TransferClient
+from globus_sdk.nexus import NexusClient
